Issue #21 - Concluding extraction; completed

+++ Extraction/Medium.h
@@ -1,4 +1,4 @@
 #ifndef MEDIUM_H
 #define MEDIUM_H
-double get_medium(double* sample, size_t n_elems);
+double get_medium(unsigned int* sample, size_t n_elems);
 #endif
